requirements.txt
pytest~=8.0.2
Appium-Python-Client~=3.2.1
python-dotenv~=1.0.1
selenium~=4.18.1